Start splitting out eval routines for Map
