+++ mathics/builtin/functional/apply_fns_to_lists.py
@@ -18,10 +18,9 @@ from mathics.core.evaluation import Evaluation
 from mathics.core.exceptions import InvalidLevelspecError, MessageException
 from mathics.core.expression import Expression
 from mathics.core.list import ListExpression
-from mathics.core.rules import is_rule
-from mathics.core.symbols import Atom, Symbol, SymbolNull, SymbolTrue
+from mathics.core.symbols import Atom, SymbolNull, SymbolTrue
 from mathics.core.systemsymbols import SymbolMapThread
-from mathics.eval.functional.apply_fns_to_lists import eval_MapAt
+from mathics.eval.functional.apply_fns_to_lists import eval_Map_level, eval_MapAt
 from mathics.eval.parts import python_levelspec, walk_levels
 from mathics.eval.patterns import param_and_option_from_optional_place
 
@@ -163,53 +162,11 @@ class Map(InfixOperator):
     def eval_level(self, f, expr, levelspec, evaluation, options={}):
         """Map[f_, expr_, Optional[levelspec_, {1}],
         OptionsPattern[Map]]"""
-
+        heads = self.get_option(options, "Heads", evaluation) is SymbolTrue
         levelspec = param_and_option_from_optional_place(
             levelspec, options, "System`Map", evaluation
         ) or ListExpression(Integer1)
-        try:
-            start, stop = python_levelspec(levelspec)
-        except InvalidLevelspecError:
-            evaluation.message("Map", "level", levelspec)
-            return
-
-        is_association = expr.has_form("Association", None)
-
-        def callback(level):
-            """
-            Map $f$ onto each element (denoted by 'level' here) at this level.
-            With exception for expr as Association, which is mapped on values only.
-            """
-            # TODO: This special behavior applies when the whole expression
-            # is of the form Association[__(Rule|RuleDelayed)], i.e., when
-            # the expression is a well-formatted Association expression.
-            # For example,
-            # `Map[F, Association[a->1,b->2, NotARule]`
-            # produces in WMA
-            # `Association[F[a->1], F[b->2], F[NotARule]`
-            # instead of
-            # `Association[a->F[1], b->F[2], F[NotARule]`]
-            #
-            # Fixing this would require a different implementation of this eval_ method.
-            #
-            if is_association and is_rule(level):
-                return Expression(
-                    level.get_head(),
-                    level.elements[0],
-                    Expression(f, level.elements[1]),
-                )
-            return Expression(f, level)
-
-        heads = self.get_option(options, "Heads", evaluation) is SymbolTrue
-        result, _ = walk_levels(expr, start, stop, heads=heads, callback=callback)
-        if isinstance(result, Symbol):
-            return result
-        elem_prop = result.elements_properties
-        if elem_prop is not None:
-            elem_prop.elements_fully_evaluated = False
-        result.elements_properties
-
-        return result
+        return eval_Map_level(f, expr, levelspec, evaluation, heads)
 
 
 class MapAt(Builtin):
